--- a/requirements.txt
+++ b/requirements.txt
@@ -6,3 +6,4 @@ rdflib==3.2.1
 argparse==1.1
 fastchardet==0.2.0
 mock==1.0.0
+simplejson==3.8.0