--- a/validator/unicodehelper.py
+++ b/validator/unicodehelper.py
@@ -1,17 +1,23 @@
 import codecs
-import textfilter
+import re
 
 # Many thanks to nmaier for inspiration and code in this module
 
-UNICODES = [
+UNICODE_BOMS = [
     (codecs.BOM_UTF8, "utf-8"),
     (codecs.BOM_UTF32_LE, "utf-32-le"),
     (codecs.BOM_UTF32_BE, "utf-32-be"),
     (codecs.BOM_UTF16_LE, "utf-16-le"),
     (codecs.BOM_UTF16_BE, "utf-16-be"),
-    ]
+]
 
-COMMON_ENCODINGS = ("utf-16", "latin_1", "ascii")
+COMMON_ENCODINGS = ("latin_1", "utf-16")
+
+# Matches any non-ASCII characters, and any unprintable characters in the
+# 7-bit ASCII range. Accepts tab, return, newline, and any other character
+# code above 0x20 which fits in 7 bits.
+NON_ASCII_FILTER = re.compile(r'[^\t\r\n\x20-\x7f]+')
+
 
 def decode(data):
     """
@@ -19,37 +25,26 @@ def decode(data):
     stripping.
     """
 
-    # Don't make more work than we have to.
-    if not isinstance(data, str):
+    if isinstance(data, unicode):
         return data
 
-    # Detect standard unicodes.
-    for bom, encoding in UNICODES:
+    # Detect standard unicode BOMs.
+    for bom, encoding in UNICODE_BOMS:
         if data.startswith(bom):
-            return unicode(data[len(bom):], encoding, "ignore")
+            return data[len(bom):].decode(encoding, errors="ignore")
 
-    # Try straight UTF-8
+    # Try straight UTF-8.
     try:
-        return unicode(data, "utf-8")
+        return data.decode("utf-8")
     except UnicodeDecodeError:
         pass
-
-    # Test for latin_1, because it can be matched as UTF-16
-    # Somewhat of a hack, but it works and is about a thousand times faster
-    # than using chardet.
-    if all(ord(c) < 256 for c in data):
-        try:
-            return unicode(data, "latin_1")
-        except UnicodeDecodeError:
-            pass
 
     # Test for various common encodings.
     for encoding in COMMON_ENCODINGS:
         try:
-            return unicode(data, encoding)
+            return data.decode(encoding)
         except UnicodeDecodeError:
             pass
 
     # Anything else gets filtered.
-    return unicode(textfilter.filter_ascii(data), errors="replace")
-
+    return NON_ASCII_FILTER.sub('', data).decode("ascii", errors="replace")