Some numbers:

Previous time to complete full run of tests: 14.6s
Current time to complete a full run of tests: 2.9s
Percent improvement: 80%

Previous time to validate https://bugzil.la/1184979: 83s
Current time to validate https://bugzil.la/1184979: 32s
Percent improvement: 61%