@kmaglione 2015-07-29
9b51091 doesn't really belong here, but I hit it after trying to time my test add-on after rebasing.