@magopian 2015-07-29
That's some really code work here, and very clever of you to tackle the low hanging fruit after having spotted them using a profiler (being curious, which tools did you use? `cProfile` and RunSnakeRun?)

I'm really surprised changing "just those" made such a significant improvement, yay!

r+wc