@kmaglione 2015-07-29
I used `cProfile` and `cprofilev` for the most part, and `line_profiler` for a few individual functions.