@kmaglione 2015-07-29
Cleaned up the, er, cleanup code, and added some tests for it. Fixed other nits.